+++ src/smartcontracts/sc_arguments.cpp
@@ -1,4 +1,4 @@
-#include "smartcontracts/scarguments.h"
+#include "smartcontracts/sc_arguments.h"
 
 #include "hex.h"
